CPU work
Added all of the cpu registers to the R3000A component, and moved the work of the parser to Instruction

+++ src/Component.cpp
@@ -1,12 +1,12 @@
 #include "Component.hpp"
 
-Component::Component(uint32 frequency):
+Component::Component(uint64 frequency):
 _frequency(frequency)
 {
     
 }
 
-uint32 Component::GetFrequency()
+uint64 Component::GetFrequency()
 {
     return _frequency;
 }
